PR updates and moved health package under util
